+++ package.json
@@ -13,6 +13,9 @@
 		"cf-typegen": "wrangler types --env-interface CloudflareEnv ./env.d.ts"
 	},
 	"dependencies": {
+		"@react-three/drei": "^10.6.1",
+		"@react-three/fiber": "^9.3.0",
+		"@react-three/postprocessing": "^3.0.4",
 		"framer-motion": "^12.23.12",
 		"next": "15.3.5",
 		"react": "^19.0.0",
@@ -20,7 +23,8 @@
 		"react-icons": "^5.5.0",
 		"react-syntax-highlighter": "^15.6.1",
 		"react-type-animation": "^3.2.0",
-		"tailwind-merge": "^3.3.1"
+		"tailwind-merge": "^3.3.1",
+		"three": "^0.179.1"
 	},
 	"devDependencies": {
 		"@cloudflare/next-on-pages": "^1.13.13",
