+++ src/components/Navbar/Navbar.tsx
@@ -1,38 +1,105 @@
 "use client";
+import { useState, useEffect } from "react";
+import Link from "next/link";
 
-import { usePathname } from "next/navigation";
-import NavItem from "./NavItem";
-import MobileMenu from "./MobileMenu";
-import { ROUTES_URL } from "@/constants";
+export default function Navbar() {
+  const [isScrolled, setIsScrolled] = useState(false);
+  const [isOpen, setIsOpen] = useState(false);
+  const [activeSection, setActiveSection] = useState("home");
 
-const Navbar = () => {
-  const pathname = usePathname();
+  const navItems = [
+    { label: "_home", href: "#home" },
+    { label: "_about", href: "#about" },
+    { label: "_projects", href: "#projects" },
+    { label: "_contact", href: "#contact" },
+  ];
+
+  // Scroll background effect
+  useEffect(() => {
+    const handleScroll = () => setIsScrolled(window.scrollY > 20);
+    window.addEventListener("scroll", handleScroll);
+    return () => window.removeEventListener("scroll", handleScroll);
+  }, []);
+
+  // Track active section
+  useEffect(() => {
+    const sections = navItems.map((item) => document.querySelector(item.href));
+
+    const observer = new IntersectionObserver(
+      (entries) => {
+        entries.forEach((entry) => {
+          if (entry.isIntersecting) {
+            setActiveSection(entry.target.id);
+          }
+        });
+      },
+      { threshold: 0.6 }
+    );
+
+    sections.forEach((sec) => sec && observer.observe(sec));
+    return () => observer.disconnect();
+  }, []);
 
   return (
-    <nav className="w-full border-b border-gray-600 shadow">
-      <div className="flex items-center justify-between p-2 md:justify-start">
-        <h1 className="pl-2 pr-12">kiran-bendkoli</h1>
+    <nav
+      className={`fixed w-full top-0 z-50 transition-all duration-300 ${
+        isScrolled ? "backdrop-blur-md bg-white/10 shadow-lg" : "bg-transparent"
+      }`}
+    >
+      <div className="max-w-6xl mx-auto px-4 py-3 flex justify-between items-center">
+        <h1 className="text-xl font-light tracking-widest  text-primary">Kiran Bendkoli</h1>
 
-        {/* Mobile menu toggle */}
-        <MobileMenu currentPath={pathname} />
+        {/* Desktop Menu */}
+        <ul className="hidden md:flex space-x-8 text-sm tracking-widest text-white">
+          {navItems.map((item) => (
+            <li key={item.href}>
+              <Link
+                href={item.href}
+                className={`transition-colors ${
+                  activeSection === item.href.slice(1)
+                    ? "text-primary border-b-2 border-primary"
+                    : "hover:text-gray-300"
+                }`}
+              >
+                {item.label}
+              </Link>
+            </li>
+          ))}
+        </ul>
 
-        {/* Desktop nav */}
-        <div className="hidden gap-4 px-4 md:flex">
-          <NavItem to={ROUTES_URL.HOME} label="_hello" activePath={pathname} />
-          <NavItem
-            to={ROUTES_URL.ABOUT}
-            label="_about-me"
-            activePath={pathname}
-          />
-          <NavItem
-            to={ROUTES_URL.PROJECTS}
-            label="_projects"
-            activePath={pathname}
-          />
-        </div>
+        {/* Mobile Hamburger */}
+        <button
+          onClick={() => setIsOpen(!isOpen)}
+          className="md:hidden flex flex-col space-y-1"
+        >
+          <span className="w-6 h-0.5 bg-white"></span>
+          <span className="w-6 h-0.5 bg-white"></span>
+          <span className="w-6 h-0.5 bg-white"></span>
+        </button>
       </div>
+
+      {/* Mobile Menu */}
+      {isOpen && (
+        <div className="md:hidden bg-black/90 backdrop-blur-md">
+          <ul className="flex flex-col items-center py-4 space-y-4">
+            {navItems.map((item) => (
+              <li key={item.href}>
+                <Link
+                  href={item.href}
+                  onClick={() => setIsOpen(false)}
+                  className={`text-lg tracking-widest ${
+                    activeSection === item.href.slice(1)
+                      ? "text-primary border-b-2 border-primary"
+                      : "text-white"
+                  }`}
+                >
+                  {item.label}
+                </Link>
+              </li>
+            ))}
+          </ul>
+        </div>
+      )}
     </nav>
   );
-};
-
-export default Navbar;
+}
