+++ src/components/StarsCanvas.tsx
@@ -1,0 +1,48 @@
+"use client";
+import { useEffect, useRef } from "react";
+
+const StarsCanvas = () => {
+  const canvasRef = useRef<HTMLCanvasElement>(null);
+
+  useEffect(() => {
+    const canvas = canvasRef.current!;
+    const ctx = canvas.getContext("2d")!;
+    canvas.width = window.innerWidth;
+    canvas.height = window.innerHeight;
+
+    const stars = Array.from({ length: 200 }, () => ({
+      x: Math.random() * canvas.width,
+      y: Math.random() * canvas.height,
+      radius: Math.random() * 1.5,
+    }));
+
+    function drawStars() {
+      ctx.clearRect(0, 0, canvas.width, canvas.height);
+      ctx.fillStyle = "white";
+      stars.forEach((star) => {
+        ctx.beginPath();
+        ctx.arc(star.x, star.y, star.radius, 0, Math.PI * 2);
+        ctx.fill();
+      });
+    }
+
+    function animate() {
+      stars.forEach((star) => {
+        star.y += 0.1;
+        if (star.y > canvas.height) star.y = 0;
+      });
+      drawStars();
+      requestAnimationFrame(animate);
+    }
+
+    animate();
+    window.addEventListener("resize", () => {
+      canvas.width = window.innerWidth;
+      canvas.height = window.innerHeight;
+    });
+  }, []);
+
+  return <canvas ref={canvasRef} className="absolute overflow-clip inset-0 z-0" />;
+}
+
+export default StarsCanvas;
